Allow a reasonable amount of cache staleness.
Add an argument to registry.get specifying a number of seconds of
staleness which is allowed in the .cache.json file before it is
re-fetched.  If the stat shows that the file hasn't changed in that much
time, then don't even bother trying to make a request, just use the data
in that file, and supply a mock response object with a 304 code.